chore(main) Estructurando

+++ src/main/java/backend/housekeeper/houseelegantservice/domain/model/aggregates/KeeperRequest.java
@@ -26,7 +26,7 @@ public class KeeperRequest extends AuditableModel {
     @Getter
     private String photoUrl;
     @Getter
-    private double rating;
+    private Float rating;
 
     @CreatedDate
     private Date createdAt;
